optimizations

+++ src/FMBot.Bot/Handlers/UserEventHandler.cs
@@ -87,7 +87,7 @@ public class UserEventHandler
     {
         Statistics.DiscordEvents.WithLabels(nameof(UserLeft)).Inc();
 
-        if (socketGuild != null && socketUser != null)
+        if (socketGuild != null && socketUser != null && PublicProperties.RegisteredUsers.ContainsKey(socketUser.Id))
         {
             _ = this._indexService.RemoveUserFromGuild(socketUser.Id, socketGuild.Id);
             _ = this._crownService.RemoveAllCrownsFromDiscordUser(socketUser.Id, socketGuild.Id);
@@ -98,7 +98,10 @@ public class UserEventHandler
     {
         Statistics.DiscordEvents.WithLabels(nameof(UserBanned)).Inc();
 
-        _ = this._indexService.RemoveUserFromGuild(guildUser.Id, guild.Id);
-        _ = this._crownService.RemoveAllCrownsFromDiscordUser(guildUser.Id, guild.Id);
+        if (guildUser != null && guild != null && PublicProperties.RegisteredUsers.ContainsKey(guildUser.Id))
+        {
+            _ = this._indexService.RemoveUserFromGuild(guildUser.Id, guild.Id);
+            _ = this._crownService.RemoveAllCrownsFromDiscordUser(guildUser.Id, guild.Id);
+        }
     }
 }
